\name{trunc_inf}
\alias{trunc_inf}
\title{Truncate to non-infinity}
\usage{
  trunc_inf(x)
}
\arguments{
  \item{x}{\code{numeric} vector}
}
\value{
  \code{numeric} vector with \code{Inf} and \code{-Inf}
  replaced with the smallest and largest floating point
  number.
}
\description{
  Truncate to non-infinity
}
\keyword{internal}

